First proposition for UnitTest framework
