+++ test/UnitTest/.gitignore
@@ -1,0 +1,20 @@
+CMakeCache.txt
+CMakeFiles/*
+CMakeLists.txt.user
+Makefile
+cmake_install.cmake
+database.cbp
+out.*
+query.dmp
+query.lkn
+query.qry*
+testfile
+x*
+CTestTestfile.cmake
+Testing/
+query.dot
+query.jpg
+doxygen.config
+*.dot
+*.jpg
+*.png
